Added Program Arguments
Added the -createOEM and -createRetail arguments, which start the application with the respective Product ID already created. These arguments both work with the -copy argument which will also copy the created Product ID to clipboard automatically.

There is also now a -notTopmost argument which will prevent the application from being always on top.